🩹 fix - BoardMember Service를 Board service로 통합
